Versioned & modularized (made optional) pubsub and bus.
In order to support future work, we want to make all the protocols
(and later the transports) explicitly version 0.  We're going to
revise many of these going forward to add missing TTLs if not already
present.

We also want to make it optional to include protocols, and make
the overrides of the default version optional too -- so that users
can supply their own overrides.  This will also facilitate building
smaller versions of the library, for applications that need only
a subset of functionality.

The tests are modified to be aware of this, and skip (or pass) as
appropriate.